Handle AUCTeX's new mode naming convention
Recently, AUCTeX renamed several of its major modes to avoid shadowing emacs built-in modes. This is documented here:
https://git.savannah.gnu.org/cgit/auctex.git/tree/doc/changes.texi

As a result, if one has the newer version of AUCTeX (the version currently available from ELPA) then when they activate LaTeX-mode, lsp will be unavailable.

This is a simple fix to add support for the new TeX-mode and LaTeX-mode.